refactor: extract useCountdownTimer hook from App.tsx
